mpegmap/HTMLSourcingInbandTracks-3.4-2.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.3-1 AudioTracks</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS - UA creates AudioTracks, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing two audio elementary streams, ES PIDs 36, 41.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>

    <audio id="a" width="400" height="300" autoplay controls></audio>
    <div id="log"></div>

    <script>
      var test = async_test("AudioTracks attributes.");
      var aud = document.getElementById("a");
      aud.src = mediaServerURL() + "ClosedCaption.ts"; 

      var expectedTracks = [
        {
          id: "36",
          kind: "main"
        }, {
          id: "41",
          kind: ""
        }
      ];

      aud.addEventListener("playing", function() {
        aud.pause();

        test.step(function() {
          assert_equals(aud.audioTracks.length, expectedTracks.length, "AudioTrackList.length");
          for (var i = 0; i < expectedTracks.length; ++i) {
            var expected = expectedTracks[i];
            // These should be in PMT order
            var actual = aud.audioTracks[i];
            assert_equals(actual.id, expected.id, "AudioTrackList[" + i + "].id");
            assert_equals(actual.kind, expected.kind, "AudioTrack[" + i + "].kind");
          }
        });
        test.done();
      }, false);
    </script>
  </body>
</html>
